User 스키마 변경
- username : email로 로그인기능을 대체
- phoneNum 삭제 : 기타 선택 개인정보는 우선은 받지 않고 최소한으로 회원가입 진행
- deleteAt 추가 : 소프트 삭제로 해당 시간이 존재하면 삭제된 데이터로 간주하게 끔 설정.
